Refactor pagination utility functions
