+++ frontend/src/pages/TodayPage.jsx
@@ -19,6 +19,7 @@ import {
 } from "@api/today.api"
 
 import { useClientTimezone } from "@utils/clientSettings"
+import { getPageFromURL } from "@utils/pagination"
 
 import queryClient from "@queries/queryClient"
 
@@ -26,14 +27,6 @@ import FeatherIcon from "feather-icons-react"
 import { DateTime } from "luxon"
 import { useTranslation } from "react-i18next"
 import { toast } from "react-toastify"
-
-const getPageFromURL = (url) => {
-    if (!url) return null
-
-    const u = new URL(url)
-    const page = u.searchParams.get("page")
-    return page
-}
 
 const TodayPage = () => {
     const { t } = useTranslation(null, { keyPrefix: "today" })
