+++ frontend/src/utils/pagination.js
@@ -1,0 +1,15 @@
+export const getPageFromURL = (url) => {
+    if (!url) return null
+
+    const u = new URL(url)
+    const page = u.searchParams.get("page")
+    return page
+}
+
+export const getCursorFromURL = (url) => {
+    if (!url) return null
+
+    const u = new URL(url)
+    const cursor = u.searchParams.get("cursor")
+    return cursor
+}
